% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p07_annotation_decoration_functions.R
\name{add_annotation_titles}
\alias{add_annotation_titles}
\title{Add annotation column titles on the heatmap}
\usage{
add_annotation_titles(annotations, anTitle, fontSize = 7)
}
\arguments{
\item{annotations}{a character vector of annotation IDs for which lables need to be added}

\item{anTitle}{a named list with titles for each annotation}

\item{fontSize}{fontsize for the annotation text}
}
\value{
The function returns no value.
}
\description{
This function use decorate_annotation function internally to add the title at the top
for each annotation
}
\examples{
NA
}
