% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_functions.R
\name{add_gene_info}
\alias{add_gene_info}
\title{Add gene information}
\usage{
add_gene_info(file, clusterDf)
}
\arguments{
\item{file}{TAB delimited file with gene information columns. Genes are listed under
column 'gene'}

\item{clusterDf}{a dataframe to which the gene information is to be added. It should
have a column 'gene'}
}
\value{
A dataframe with various gene information columns added
}
\description{
Add gene information
}
\examples{
NA
}
