% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_data_import_functions.R
\name{add_SM_cluster_info}
\alias{add_SM_cluster_info}
\title{Title}
\usage{
add_SM_cluster_info(file, clusterDf)
}
\arguments{
\item{file}{file}

\item{clusterDf}{dataframe}
}
\value{
A dataframe
}
\description{
Title
}
\examples{
NA
}
