% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{annotate_ranges}
\alias{annotate_ranges}
\title{Annotate GRanges using TxDB}
\usage{
annotate_ranges(peaks, txdb, promoterLength, txIds = NULL,
  blacklistRegions = NULL, excludeType = c("tRNA", "rRNA", "snRNA",
  "snoRNA", "ncRNA"), includeFractionCut = 0.7, bindingInGene = FALSE,
  insideSkewToEndCut = 0.7, removePseudo = FALSE)
}
\arguments{
\item{peaks}{A GRanges object with name column.}

\item{txdb}{TxDB object which will be used for annotation}

\item{promoterLength}{Promoter length in number of nucleotides.}

\item{txIds}{A vector of transcript IDs to be used specifically in the annotation
process instead of full transcript set. These should be internal tx_ids from TxDB
object. This is useful feature to exclude tRNA, rRNA transcripts while annotating
the regions. Default: NULL}

\item{blacklistRegions}{A BED file or GRanges object with ChIPseq blacklist regions.
Peaks overlapping with these regions are not used for annotation.}

\item{excludeType}{Types of transcripts to exclude from annotation. Should be a
character vector. Default: \code{c("tRNA", "rRNA", "snRNA", "snoRNA", "ncRNA")}}

\item{includeFractionCut}{Number between [0, 1]. If a peak covers more than this
fraction of feature/gene, it will be marked as include_tx/include_CDS. Default: 0.7}

\item{bindingInGene}{Logical: whether the ChIPseq TF binds in gene body. This is
useful for polII ChIPseq data. Default: FALSE}

\item{insideSkewToEndCut}{A floating point number in range [0, 1]. If a peak is
present inside feature/gene and the relative summit position is > insideSkewToEndCut,
it is closer to the end of the feature. Default: 0.7}

\item{removePseudo}{Logical: whether to remove peak targets which are marked as pseudo.
Default: FALSE}
}
\value{
A GenomicRanges object with peak annotation
}
\description{
This is a generic function to annotate GRanges object with TxDB annotations. \cr
Regions are annnotated with following \strong{broad categories} and \emph{specific
types} (listed in decreasing order of preference):
\enumerate{
\item \strong{featureInPeak:} \emph{"include_tx", "include_CDS"}
\item \strong{nearStart:} \emph{"5UTR", "CDS_start", "tx_start"}
\item \strong{nearEnd:} \emph{"3UTR", "tx_end", "CDS_end"}
\item \strong{peakInFeature:} \emph{"exon", "intron", "inside_tx", "inside_CDS"}
\item \strong{upstreamTss:} \emph{"promoter", "upstream"}
\item \strong{intergenic:} \emph{"intergenic"}
}
Additionally, a \emph{pseudo} prefix is added to the peakType where a peak is
annotated to two target genes/features and one of it is more optimum than other.
The less optimum target type is prefixed with \emph{pseudo}.
}
\examples{
NA
}
