% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{annotate_regions}
\alias{annotate_regions}
\title{annotate regions using TxDB object}
\usage{
annotate_regions(peaks, txdb, txIds = NULL, blacklistRegions = NULL,
  excludeType = c("tRNA", "rRNA", "snRNA", "snoRNA", "ncRNA"),
  includeFractionCut = 0.7, bindingInGene = FALSE, promoterLength,
  insideSkewToEndCut = 0.7, removePseudo = FALSE, output = NULL)
}
\arguments{
\item{peaks}{A GRanges object with name column.}

\item{txdb}{TxDB object which will be used for annotation}

\item{txIds}{A vector of transcript IDs to be used specifically in the annotation
process instead of full transcript set. These should be internal tx_ids from TxDB
object. This is useful feature to exclude tRNA, rRNA transcripts. Default: NULL}

\item{blacklistRegions}{A BED file with ChIPseq blacklist regions. Peaks overlapping
with these regions are not used for annotation.}

\item{excludeType}{Types of transcripts to exclude from annotation. Should be a
character vector. Default: \code{c("tRNA", "rRNA", "snRNA", "snoRNA", "ncRNA")}}

\item{includeFractionCut}{Number between [0, 1]. If a peak covers more than this
fraction of feature/gene, it will be marked as include_tx/include_CDS. Default: 0.7}

\item{bindingInGene}{Logical: whether the ChIPseq TF binds in gene body. This is
useful for polII ChIPseq data. Default: FALSE}

\item{promoterLength}{Promoter length in number of nucleotides.}

\item{insideSkewToEndCut}{A floating point number in range [0, 1]. If a peak is
present inside feature/gene and the relative summit position is > insideSkewToEndCut,
it is closer to the end of the feature. Default: 0.7}

\item{removePseudo}{Logical: whether to remove peak targets which are marked as pseudo.
Default: FALSE}

\item{output}{Optionally store the annotation output to a file}
}
\value{
A GenomicRanges object with peak annotation
}
\description{
annotate regions using TxDB object
}
\examples{
NA
}
