% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_replicate_peakset.R
\name{best_replicate_peakset}
\alias{best_replicate_peakset}
\title{Get good replicate's peakset from two or more ChIPseq replicates}
\usage{
best_replicate_peakset(sampleInfo, cdsFile, ...)
}
\arguments{
\item{sampleInfo}{Sample information dataframe}

\item{cdsFile}{BED file with gene information}

\item{...}{Other arguments for \code{preProcess_macs2_results()} function}
}
\value{
A dataframe with target gene information for confident peaks
}
\description{
This function internally uses \code{chipmine::combinatorial_binding_matrix}
function to generate a peak overlap matrix and selects the peaks which are
present in all the replicates. Then it uses the best replicate to extract the
target gene infomation.
}
\examples{
NA
}
