% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p03_QC_replicate_correlation.R
\name{compare_ChIPseq_replicates}
\alias{compare_ChIPseq_replicates}
\title{TF ChIPseq replicate correlation plots}
\usage{
compare_ChIPseq_replicates(sampleInfo, compare = "pvalue",
  yintercept = -Inf, summitRegion = 0)
}
\arguments{
\item{sampleInfo}{sample information dataframe}

\item{compare}{One of \emph{pvalue} or \emph{enrichment}}

\item{yintercept}{yintercept for horizontal line on beeswarm plot}

\item{summitRegion}{Region width around peak summit to use for while merging the peaks
from multiple samples. With increasing number of peaksets, the mearging of peaks creates
broader consensus peaksets. Using a small region around peak summit allows to limit this
consensus peak width. If 0 (default), whole peak region is used. If \code{summitRegion > 0},
\code{2 x summitRegion} region around peak summit is used to create consensus peakset.}
}
\value{
A list with following ggplot2 elements:
\itemize{
\item \strong{table:} peak overlap summary table for replicates as a \code{ggtable} object
\item \strong{distribution:} Beeswarm plots showing comparative distribution between
replicates for a metric of choice
\item \strong{valueScatter:} X-Y scatter plot using values a metric of choice
\item \strong{rankScatter:} X-Y scatter plot using \code{rank(values)} a metric of choice
}
}
\description{
This function generate the peak overlap statistics table for two TF ChIPseq
replicates. Using common peaks, it plots the distribution and scatter plot of
either -log10(p-value) or fold enrichment from peak file. Macs2 peak caller
output files in narrowPeak/broadPeak format are used to extract the values.
}
\examples{
NA
}
