% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p03_data_preprocess_functions.R
\name{gene_level_peak_annotation}
\alias{gene_level_peak_annotation}
\title{Pre-process peak annotation data}
\usage{
gene_level_peak_annotation(sampleId, peakAnnotation, genesDf, peakFile,
  bwFile, preference = c("nearStart", "peakInFeature", "featureInPeak",
  "nearEnd", "upstreamTss", "intergenic"), outFile = NULL,
  bindingInGene = FALSE, removePseudo = TRUE, fcCutoff = 1,
  pvalCutoff = 1)
}
\arguments{
\item{sampleId}{sample ID}

\item{peakAnnotation}{A file path or a dataframe.
The file should be macs2 narrowPeak file annotated with function
\code{narrowPeak_annotate()}.
Dataframe should be generated by \code{import_peak_annotation()} function. A subset of
rows can be provided for special downstream usage.}

\item{genesDf}{A dataframe with genes and other information. There must be a geneId
column.}

\item{peakFile}{macs2 output file. Should be narrowPeak or broadPeak file}

\item{bwFile}{bigWig format coverage track file}

\item{preference}{A character vector with 5 values: \code{c("nearStart", "peakInFeature",
"featureInPeak", "nearEnd", "upstreamTss", "intergenic")}. If the order of the values is
used as preference order for selecting the most appropriate peak for a gene.
 Default: \code{c("nearStart", "peakInFeature", "featureInPeak", "nearEnd", "upstreamTss")}}

\item{outFile}{Optional output file. Default: NULL which means the output is not stored}

\item{bindingInGene}{whether the factor bind in gene body like polII or not}

\item{removePseudo}{Logical: whether to remove peak targets which are marked as pseudo.
Default: TRUE}

\item{fcCutoff}{fold enrichment cutoff for the peak. Default: 1 i.e. no cutoff}

\item{pvalCutoff}{log10_pvalue cutoff for the peak. Default: 1 i.e. no cutoff}
}
\value{
Returns a dataframe with processed peak annotation information
}
\description{
Main purpose of this function is to create a table of gene level peak annotation.\cr
This function processes the peak annotation generated by function
\code{narrowPeak_annotate()}. If a gene has multiple peaks of category \emph{"nearEnd"} or
\emph{"peakInFeature"}, peak with highest significance is selected. Finally,
\code{preference} and \code{peakPval} is used to select best peak in case there are more
than one peaks annotated with same gene. \cr
\emph{"intergenic"} peaks are not reported here as they are not annotated with gene
information. \cr
If user want annotations for all peaks, \code{import_peak_annotation()} function
should be used instead.
}
\examples{
NA
}
