% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p08_average_profile_functions.R
\name{geneset_average_profile}
\alias{geneset_average_profile}
\title{Calculate average profile for a gene list}
\usage{
geneset_average_profile(exptInfo, profileMats, genes,
  cluster = "group_A")
}
\arguments{
\item{exptInfo}{Experiment information dataframe}

\item{profileMats}{a named list of profile matrix.}

\item{genes}{a vector of gene IDs. Profile matrix will be subset based on this
gene list to calculate average profile}

\item{cluster}{A group name for current profile. Default: group_A}
}
\value{
A dataframe which can be used for plotting average profile
}
\description{
This method calculates the average profile for gene of interest.
}
\examples{
NULL
}
