% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_multiTF_plot.R
\name{geneset_multiTF_plot}
\alias{geneset_multiTF_plot}
\title{Title}
\usage{
geneset_multiTF_plot(data, exptInfo, outFile, plotTitle, name, res,
  numClust)
}
\arguments{
\item{numClust}{}
}
\description{
Title
}
\examples{
NA
}
