% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot_list_functions.R
\name{get_expression_heatmap_list}
\alias{get_expression_heatmap_list}
\title{Multiple PolII expression heatmaps as list}
\usage{
get_expression_heatmap_list(expDf, exptInfo, genes, htColor = NULL)
}
\arguments{
\item{expDf}{a dataframe with PolII expression columns}

\item{exptInfo}{experiment info as data frame with information like sampleID, type, path etc}

\item{genes}{A vector of gene IDs which are to be plotted}

\item{htColor}{heatmap color generated by \code{ColorRamp2}}
}
\value{
A named list object with multiple Heatmaps
}
\description{
Multiple PolII expression heatmaps as list
}
\examples{
NA
}
