% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_fold_change_functions.R
\name{get_fold_change}
\alias{get_fold_change}
\title{Raw log2 fold change value between two samples as log2(s2/s1)}
\usage{
get_fold_change(df, nmt, dmt, newCol = "lfc", isExpressedCols = NULL)
}
\arguments{
\item{df}{data frame}

\item{nmt}{column name for sample in numerator}

\item{dmt}{column name for sample in denominator}

\item{newCol}{new column name for log2(nmt/dmt) values. Default: lfc}

\item{isExpressedCols}{A named vector with column names for isExpressed data for
both polII samples}
}
\value{
dataframe with additional fold change column
}
\description{
Raw log2 fold change value between two samples as log2(s2/s1)
}
\examples{
NA
}
