% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p03_get_narrowpeak_summit_seq.R
\name{get_narrowpeak_summit_seq}
\alias{get_narrowpeak_summit_seq}
\title{Get DNA sequence from peak summit region}
\usage{
get_narrowpeak_summit_seq(npFile, id, genome, length = 200,
  columnNameWithIdPrefix = TRUE)
}
\arguments{
\item{npFile}{narrowpeak file generated by macs2}

\item{id}{sample ID}

\item{genome}{BSgenome object}

\item{length}{Length of the region. Default: 200}

\item{columnNameWithIdPrefix}{Whether to modify the column names to include sampleId.
If TRUE, the new column names are of format: sampleId(colName)}
}
\value{
A dataframe with three columns: name, summitRegion, summitSeq
}
\description{
This function extract the DNA sequence around peak summit region from narrowPeak file
}
\examples{
summitSeq = get_narrowpeak_summit_seq("test.narropeak", "sampleId", BSgenome.Afumigatus.AspGD.Af293, 200)

}
