% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_annotation_args.R
\name{get_peak_annotation_args}
\alias{get_peak_annotation_args}
\title{Get peak heatmap annotation}
\usage{
get_peak_annotation_args(tssCol, tesCol)
}
\arguments{
\item{tssCol}{column name of the TSS region peaks}

\item{tesCol}{column name of the TES region peaks}
}
\value{
A list of two elements: peak type color and peak type legend
}
\description{
Get peak heatmap annotation
}
\examples{
NA
}
