% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_data_import_functions.R
\name{get_peak_region_expression}
\alias{get_peak_region_expression}
\title{Add peak region RPM score}
\usage{
get_peak_region_expression(exptInfo, genesDf)
}
\arguments{
\item{exptInfo}{sample information dataframe generated by function get_sample_information()}

\item{genesDf}{a dataframe to which the peak region RMP score column will be added}
}
\value{
A data frame
}
\description{
Add peak region RPM score
}
\examples{
NA
}
