% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p03_get_narrowpeak_summit_seq.R
\name{get_peak_summit_seq}
\alias{get_peak_summit_seq}
\title{Get DNA sequence from peak summit region}
\usage{
get_peak_summit_seq(file, peakFormat = "narrowPeak", sampleId, genome,
  length = 200, column_name_prefix = TRUE)
}
\arguments{
\item{file}{peak file generated by macs2}

\item{peakFormat}{Peak file format. One of \code{c("narrowPeak", "broadPeak", "bed")}
Default: narrowPeak}

\item{sampleId}{sample ID}

\item{genome}{BSgenome object}

\item{length}{Length of the region. Default: 200}

\item{column_name_prefix}{Whether to modify the column names to include sampleId.
If TRUE, the new column names are of format: sampleId.colName}
}
\value{
A dataframe with three columns: name, summitRegion, summitSeq
}
\description{
This function extract the DNA sequence around peak summit region from narrowPeak file
}
\examples{
summitSeq = get_peak_summit_seq("test.narropeak", "sampleId", BSgenome.Afumigatus.AspGD.Af293, 200)

}
