% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_data_import_functions.R
\name{get_polII_signal}
\alias{get_polII_signal}
\title{Read polII binding signal data for a sample}
\usage{
get_polII_signal(file, title, clusterData)
}
\arguments{
\item{file}{polII binding signal file}

\item{title}{Title of the polII sample}

\item{clusterData}{A dataframe to which polII binding signal column should be added.
This dataframe should have 'geneId' column}
}
\value{
A list with three elements:
\itemize{
\item log2_mat: a matrix with log2(polII binding signal) values for heatmap generation
\item clusterDf: dataframe with polII binding signal columns
\item quantiles: polII binding signal quantile values generated by quantile() function
}
}
\description{
This function reads the polII binding data for single sample and add it
to the provided dataframe. It also convert the polII signal values to log2
which can be used for plotting polII signal heatmap
}
\examples{
NA
}
