% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot_list_functions.R
\name{get_profile_plot_list}
\alias{get_profile_plot_list}
\title{Multiple Enriched Heatmaps as list}
\usage{
get_profile_plot_list(exptInfo, cluster = NULL, clusterColor = NULL,
  clustOrd = NULL, geneList, colorList = NULL,
  matrixSource = "normalizedmatrix", matrixBins = c(200, 200, 100, 10),
  targetType = "region", targetName = "gene", ylimFraction = NULL,
  keep = c(0, 1), rasterPar = list(use = TRUE, qual = 5), ...)
}
\arguments{
\item{exptInfo}{experiment info as data frame with information like sampleID, type,
path etc}

\item{cluster}{Dataframe with cluster information. Two columns must be present
in this dataframe: 'cluster', 'geneId'. Default: NULL}

\item{clusterColor}{cluster annotation color information. Default: NULL}

\item{clustOrd}{A character vector of cluster order in the plot. If not provided,
clusters are arranged as per character sort order}

\item{geneList}{A vector of gene IDs which are to be plotted}

\item{colorList}{a named list of color objects for each of the sample. Color for
profile heatmap should be generated by colorRamp2}

\item{matrixSource}{Method by which profile matrix was generated. One of \emph{deeptools, miao,
normalizedmatrix}. Default: normalizedmatrix}

\item{matrixBins}{A numeric vector with four elements representing the column
counts for up, body, down regions and binSize for the profile matrix.
Default: \code{c(200, 200, 100, 10)}}

\item{targetType}{One of "region", "point". If targetType is "region", target is used to decide
the number of bins over region. Otherwise, for "point", a signle point matrix is expected where
the region is extracted around single point. Default: region}

\item{targetName}{Name to be used for target. Eg: "gene", "TSS", "TES", "summit" etc. default: gene}

\item{ylimFraction}{A named list with intensity scores to use as ylimit for top annotation.
If the value is single number, it has to be floating point number to extract the quantile and use
limit \code{[0, quantile(x)]}. If the value is numeric vector of length 2, the two numbers are used as lower
and upper limits for ylimit of top annotation. Default: NULL.}

\item{keep}{Same as \code{EnrichedHeatmap::normalizeToMatrix}. First value is used as lower quantile
and any value in profile matrix less than lower quantile is set to lower quantile. Second value is
used as upper quantile and any value greater than upper quantile is set to upper quantile.
Default: \code{c(0, 1)}}

\item{rasterPar}{rasterization parameters as list() with two elements: use, qual. Default:
\code{list(use = TRUE, qual = 5)}. \code{rasterPar$use} is used for \code{use_raster} argument and
\code{rasterPar$qual} is used for \code{raster_quality} argument of \code{Heatmap} function.}

\item{...}{Other arguments for \code{EnrichedHeatmap} function}
}
\value{
A named list of Enriched Heatmap objects
}
\description{
Multiple Enriched Heatmaps as list
}
\examples{
NA
}
