% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{get_TES}
\alias{get_TES}
\title{Extract the TES position as GenomicRanges}
\usage{
get_TES(gr, up = 0, down = 1, ...)
}
\arguments{
\item{gr}{A GenomicRange object}

\item{up}{upstream region length. This is 'upstream' argument of  GenomicRanges::promoters
method}

\item{down}{downstream region length. This is 'downstream' argument of  GenomicRanges::promoters
method}

\item{...}{Other arguments to GenomicRanges::promoters function}
}
\value{
A GenomicRange object with TES position
}
\description{
This function extract the TES regions from the GenomicRanges. Internally, it invert the strand
information of GenomicRanges and then use the promoters() method to extract the promoter
region for this flipped ranges. Finally the strand information is switched back again.
}
\examples{
NULL
}
