% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_TF_binding_data.R
\name{get_TF_binding_data}
\alias{get_TF_binding_data}
\title{Get the TF peak annotation data}
\usage{
get_TF_binding_data(genesDf, exptInfo, allColumns = FALSE)
}
\arguments{
\item{genesDf}{Dataframe with gene of interest}

\item{exptInfo}{Experiment information dataframe generated by get_sample_information()}

\item{allColumns}{Whether to get all the columns TRUE or FALSE}
}
\value{
Dataframe with all columns when allColumns is TRUE.
If allColumns = FALSE, "hasPeak.sampleId",  "peakId.sampleId",  "peakDist.sampleId",
"summitDist.sampleId",  "peakType.sampleId",   "enrichment.sampleId",  "pval.sampleId",
"tesPeakType.sampleId",  "tesEnrichment.sampleId",  "tesPval.sampleId"  columns are
returned from TF peak annotation file
}
\description{
This function extract the columns from SAMPLE_ID_peaks.annotated.tab file for
 multiple samples and returns a dataframe.
}
\examples{
NA
}
