% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p99_utils_txdb.R
\name{get_txdb_fiveUtr_gr}
\alias{get_txdb_fiveUtr_gr}
\title{5UTR GRanges object from TxDB}
\usage{
get_txdb_fiveUtr_gr(txdb, txIds = NULL)
}
\arguments{
\item{txdb}{\code{TxDB} object which will be used for annotation}

\item{txIds}{A vector of transcript IDs to be used specifically in the annotation
process instead of full transcript set. These should be internal tx_ids from \code{TxDB}
object. This is useful feature to exclude tRNA, rRNA transcripts while annotating
the regions. Default: NULL}
}
\value{
GRanges object from TxDB
}
\description{
5UTR GRanges object from TxDB
}
\examples{
NA
}
