% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p99_utils_txdb.R
\name{get_txdb_transcripts_gr}
\alias{get_txdb_transcripts_gr}
\title{Transcripts GRanges object from TxDB}
\usage{
get_txdb_transcripts_gr(txdb, excludeType = NULL, txIds = NULL)
}
\arguments{
\item{txdb}{\code{TxDB} object which will be used for annotation}

\item{excludeType}{Types of transcripts to exclude from annotation. Should be a
character vector. Default: \code{c("tRNA", "rRNA", "snRNA", "snoRNA", "ncRNA")}.
This argument work only when \code{TxDB} object has \code{TXTYPE} column with
appropriate transcripy type values.}

\item{txIds}{A vector of transcript IDs to be used specifically in the annotation
process instead of full transcript set. These should be internal tx_ids from \code{TxDB}
object. This is useful feature to exclude tRNA, rRNA transcripts while annotating
the regions. Default: NULL}
}
\value{
GRanges object from TxDB
}
\description{
Transcripts GRanges object from TxDB
}
\examples{
NA
}
