% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_data_import_functions.R
\name{get_upstream_expression}
\alias{get_upstream_expression}
\title{Add 500bp upstream region RPM score}
\usage{
get_upstream_expression(exptInfo, genesDf)
}
\arguments{
\item{exptInfo}{sample information dataframe generated by function get_sample_information()}

\item{genesDf}{a dataframe to which the peak region RPM score column will be added}
}
\value{
A data frame
}
\description{
Add 500bp upstream region RPM score
}
\examples{
NA
}
