% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p03_data_import_functions.R
\name{import_peaks_as_df}
\alias{import_peaks_as_df}
\title{Read peak file as dataframe with selected columns}
\usage{
import_peaks_as_df(
  file,
  peakFormat = "narrowPeak",
  peakCols = c("peakId", "peakEnrichment", "peakPval"),
  sampleId = NULL,
  column_suffix = TRUE
)
}
\arguments{
\item{file}{Peak file}

\item{peakFormat}{one of \code{c("narrowPeak", "broadPeak", "bed")}}

\item{peakCols}{Column to select in dataframe. Column names should be from this list:
\code{c("peakChr", "peakStart", "peakEnd", "peakId", "peakScore", "peakStrand",
"peakEnrichment", "peakPval", "peakQval", "peakSummit")}.
Default: \code{c("peakId", "peakEnrichment", "peakPval")}}

\item{sampleId}{sampleId to be used as suffix in column names}

\item{column_suffix}{Logical: whether to add sampleId suffix to column name. Default: TRUE}
}
\value{
A dataframe with columns specified
}
\description{
Read peak file as dataframe with selected columns
}
\examples{
NA
}
