% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p05_profile_matrix_import_functions.R
\name{import_profiles}
\alias{import_profiles}
\title{Generate profile matrix list}
\usage{
import_profiles(exptInfo, geneList, ...)
}
\arguments{
\item{exptInfo}{experiment info as data frame with information like sampleID, type, path etc}

\item{geneList}{A vector of gene IDs for which the profile needs to be extracted}

\item{...}{Other argument to \code{import_profile_from_file} function}
}
\value{
A list of normalized matrix where each matrix is of \code{normalizedMatrix} class
}
\description{
This function reads/generate the profile matrix of \code{normalizedMatrix} class for each
sample
}
\examples{
NA
}
