% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p05_merged_profile_matrix_cluster.R
\name{merged_profile_matrix_cluster}
\alias{merged_profile_matrix_cluster}
\title{Merged profile matrix clustering}
\usage{
merged_profile_matrix_cluster(name, exptInfo, genes, clusterStorePath,
  matrixBins = c(200, 200, 100, 10), source = "deeptools", k,
  kmIter = 200, kmStarts = 200, ...)
}
\arguments{
\item{name}{name of the newly mearged profile matrix}

\item{exptInfo}{experiment info as data frame with information like sampleID, type, path etc}

\item{genes}{A vector of gene IDs which are to be plotted}

\item{clusterStorePath}{Path for storing/accessing the cluster information file.}

\item{matrixBins}{A numeric vector with four elements representing the column
counts for up, body, down regions and binSize for the profile matrix.
Default: c(200, 200, 100, 10)}

\item{source}{Method by which profile matrix was generated.
One of "deeptools", "miao" or "normalizedmatrix".}

\item{k}{number of cluster centers}

\item{kmIter}{iter.max argument of stats::kmeans() function. Default: 200}

\item{kmStarts}{nstart argument of stats::kmeans() function. Default: 200}

\item{...}{other arguments for stats::kmeans() function}

\item{matrixBins}{A numeric vector with four elements}
}
\value{
A list with two elements:
\itemize{
\item profileKm: kmeans clustering object returned by function profile_matrix_kmeans()
\item mat: a merged profile matrix as dataframe
}
}
\description{
This function join the multiple profile matrix and performs the k-means clustering
on the merged profile matrix
}
\examples{
NA
}
