% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p05_profile_matrix_import_functions.R
\name{na_impute}
\alias{na_impute}
\title{Replace NA values by row mean}
\usage{
na_impute(index, mat)
}
\arguments{
\item{index}{row number with NA values found}

\item{mat}{matrix}
}
\value{
corrected row in which NA values are replaced by mean of row
}
\description{
Replace NA values by row mean
}
\examples{
NA
}
