% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{narrowPeak_annotate}
\alias{narrowPeak_annotate}
\title{Annotate narrowPeak using TxDB}
\usage{
narrowPeak_annotate(peakFile, fileFormat = "narrowPeak",
  summitRegion = 0, txdb, promoterLength, txIds = NULL,
  blacklistRegions = NULL, excludeType = c("tRNA", "rRNA", "snRNA",
  "snoRNA", "ncRNA"), includeFractionCut = 0.7, bindingInGene = FALSE,
  insideSkewToEndCut = 0.7, removePseudo = FALSE, output = NULL)
}
\arguments{
\item{peakFile}{A narroPeak or broadPeak file. If a broadPeak file, peak center is
used as summit as broadPeak file does not report summit}

\item{fileFormat}{Format of the peak file. One of "narrowPeak" (Default) or "broadPeak".}

\item{summitRegion}{Region width around peak summit to use for annotation purpose. This
allows peaks with uniform peak width centered around summit. If 0, whole peak region
is used. If > 0, it indicates how many basepairs to include upstream and downstream
of the peak summit.}

\item{txdb}{TxDB object which will be used for annotation}

\item{promoterLength}{Promoter length in number of nucleotides.}

\item{txIds}{A vector of transcript IDs to be used specifically in the annotation
process instead of full transcript set. These should be internal tx_ids from TxDB
object. This is useful feature to exclude tRNA, rRNA transcripts while annotating
the regions. Default: NULL}

\item{blacklistRegions}{A BED file or GRanges object with ChIPseq blacklist regions.
Peaks overlapping with these regions are not used for annotation.}

\item{excludeType}{Types of transcripts to exclude from annotation. Should be a
character vector. Default: \code{c("tRNA", "rRNA", "snRNA", "snoRNA", "ncRNA")}}

\item{includeFractionCut}{Number between [0, 1]. If a peak covers more than this
fraction of feature/gene, it will be marked as include_tx/include_CDS. Default: 0.7}

\item{bindingInGene}{Logical: whether the ChIPseq TF binds in gene body. This is
useful for polII ChIPseq data. Default: FALSE}

\item{insideSkewToEndCut}{A floating point number in range [0, 1]. If a peak is
present inside feature/gene and the relative summit position is > insideSkewToEndCut,
it is closer to the end of the feature. Default: 0.7}

\item{removePseudo}{Logical: whether to remove peak targets which are marked as pseudo.
Default: FALSE}

\item{output}{Optionally store the annotation output to a file}
}
\value{
A GenomicRanges object with peak annotation
}
\description{
This function annotate the MACS2 called peaks with appropriate target transcript and
gene from TxDB object. Peaks are annnotated with following \strong{broad categories} and \emph{specific
types} (listed in decreasing order of preference):
\enumerate{
\item \strong{featureInPeak:} \emph{"include_tx", "include_CDS"}
\item \strong{nearStart:} \emph{"5UTR", "CDS_start", "tx_start"}
\item \strong{nearEnd:} \emph{"3UTR", "tx_end", "CDS_end"}
\item \strong{peakInFeature:} \emph{"exon", "intron", "inside_tx", "inside_CDS"}
\item \strong{upstreamTss:} \emph{"promoter", "upstream"}
\item \strong{intergenic:} \emph{"intergenic"}
}
Additionally, a \emph{pseudo} prefix is added to the peakType where a peak is
annotated to two target genes/features and one of it is more optimum than other.
The less optimum target type is prefixed with \emph{pseudo}. Please refer to the
\strong{Details} section for specific information on this.
}
\details{
Some important observations to do before annotating ChIPseq data:
\enumerate{
\item Whether the signal is like TF/polII i.e. factor binds across whole gene or not.
Also see if binding is throughout the genome like CTCF factor.
See \code{bindingInGene, promoterLength} arguments for the details.
\item For the genes which are within peak region, what is the gene size (are
genes shorter in length than normal) and how far is the next downstream gene.
See \code{includeFractionCut} argument for the details.
\item Are there any TES or 3' UTR peaks and how confident are they?
\item Check the TXTYPE in TxDB object and see which type of features are of
interest to you. Usually tRNA, rRNA are not needed. See \code{excludeType}
argument for the details.
}
These observations will help to decide appropriate parameters while annotating
the peaks using TxDB object.
}
\examples{
NA
}
