% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{nearest_upstream_bidirectional}
\alias{nearest_upstream_bidirectional}
\title{True target for bidirectional peak}
\usage{
nearest_upstream_bidirectional(bdirTargets, skewFraction = 0.2,
  minTSS_gapForPseudo = 500)
}
\arguments{
\item{bdirTargets}{A dataframe with two rows for bidirectional targets}

\item{skewFraction}{Minimum fraction of peak region allowed on the side of false target
from the midpoint of two target genes. Default: 0.2}

\item{minTSS_gapForPseudo}{Valid distance between two target genes TSS to mark one as psuedo}
}
\value{
Same dataframe in which peakType for one of the target is marked as pseudo
}
\description{
This function
}
\details{
target1              peak                         target2

==<=====<=====<===     -------             ===>=====>=====>=====>==

|
                   center between two targets

True target: target1: more than 80% of the peak lies on target1 side
}
\examples{
NULL
}
