% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_target_matrix.R
\name{peak_target_matrix}
\alias{peak_target_matrix}
\title{Peak target gene matrix for multiple samples}
\usage{
peak_target_matrix(sampleInfo, position = "TSS")
}
\arguments{
\item{sampleInfo}{sample information dataframe}

\item{position}{One of \code{c("TSS", "TES", "best", "both")}. This is used
for filtering the target genes for each sample. Default: "TSS". Refer to Details
section for more information.}
}
\value{
A gene wise peak target dataframe for multiple samples
}
\description{
This function generates gene based peak target matrix for multiple samples. If
\code{sampleInfo} has just one row for a sample, this function will return
per gene peak target for just one sample. If there are multiple samples,
a full join is performed to extract peaks in all samples for each gene.
}
\details{
A gene can have two types of peaks, one at TSS and another at TES. \code{position}
option allows to select the peak target in following manner.
\itemize{
\item \strong{TSS:} select genes with peak signal upstream or near TSS
\item \strong{TES:} select genes with peak signal near end of the gene
\item \strong{best:} If a gene have both TSS and TES region peak, select best based
on log10Pvalue of peak caller
\item \strong{both:} If a gene have both TSS and TES region peak, select both peaks
}
}
\examples{
NA
}
