% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_data_preprocess_functions.R
\name{peak_targets_at_TSS}
\alias{peak_targets_at_TSS}
\title{Extract genes from peak annotation data where binding is near TSS}
\usage{
peak_targets_at_TSS(sampleId, peakAnotation)
}
\arguments{
\item{sampleId}{sample ID}

\item{peakAnotation}{a dataframe containing peak annotation information generated by
function \code{import_peak_annotation}}
}
\value{
A TSS target gene dataframe with sampleId added as suffix to column names
}
\description{
Extract genes from peak annotation data where binding is near TSS
}
\examples{
NA
}
