% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p03_distribution_plot_functions.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Generate XY scatter plot for two sets}
\usage{
plot_scatter(df, sx, sy, title = "Scatter plot", colorCol,
  transformation = "identity", pseudoCount = 0)
}
\arguments{
\item{df}{dataframe with values two columns for which scatter plot is to be generated}

\item{sx}{sample 1 column name}

\item{sy}{sample 2 column name}

\item{title}{title to be showed on plot}

\item{colorCol}{column name which should be used for coloring the points}

\item{transformation}{trans argument from ggplot::scale_x_continuous() function.
Default: identity}

\item{pseudoCount}{A small number to be added to the values to avoid log(0) error.
Default: 0.1}
}
\value{
a ggplot object
}
\description{
Generate XY scatter plot for two sets
}
\examples{
NA
}
