% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_polII_exploratory_functions.R
\name{polII_tf_expression_scatter}
\alias{polII_tf_expression_scatter}
\title{Title}
\usage{
polII_tf_expression_scatter(df, tf, polII, tfExp, polIIExp, isExpCol,
  isBoundCol)
}
\arguments{
\item{isBoundCol}{}
}
\description{
Title
}
