% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess_functions.R
\name{preProcess_macs2_results}
\alias{preProcess_macs2_results}
\title{Pre-process peak annotation data}
\usage{
preProcess_macs2_results(sampleId, peakAnnotation, cdsFile, peakFile,
  bwFile, outFile = NULL, bindingInGene = FALSE, fcCutoff = 1,
  pvalCutoff = 1)
}
\arguments{
\item{sampleId}{sample ID}

\item{peakAnnotation}{A file path or a dataframe.
The file should be macs2 narrowPeak file annotated with function
\code{narrowPeak_annotate()}.
Dataframe should be generated by \code{import_peak_annotation()} function. A subset of
rows can be provided for special downstream usage.}

\item{cdsFile}{BED file with gene information}

\item{peakFile}{macs2 output file. Should be narrowPeak or broadPeak file}

\item{bwFile}{bigWig format coverage track file}

\item{outFile}{Optional output file. Default: NULL which means the output is not stored}

\item{bindingInGene}{whether the factor bind in gene body like polII or not}

\item{fcCutoff}{fold enrichment cutoff for the peak. Default: 1 i.e. no cutoff}

\item{pvalCutoff}{log10_pvalue cutoff for the peak. Default: 1 i.e. no cutoff}
}
\value{
Returns a dataframe with processed peak annotation information
}
\description{
This function processes the peak annotation generated for narrowPeak file by
function \code{narrowPeak_annotate()}. It prepares a gene level ChIPseq signal
file.
}
\examples{
NA
}
