% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_matrix_kmeans.R
\name{profile_matrix_kmeans}
\alias{profile_matrix_kmeans}
\title{Profile Matrix k-means Clustering}
\usage{
profile_matrix_kmeans(mat, km, clustFile, name, kmIter = 100,
  kmStarts = 150, ...)
}
\arguments{
\item{mat}{profile matrix}

\item{km}{number of cluster centers}

\item{kmIter}{iter.max    argument of stats::kmeans() function. Default: 100}

\item{kmStarts}{nstart argument of stats::kmeans() function. Default: 150}

\item{...}{other arguments for stats::kmeans() function}
}
\value{
A list with two elements:
\itemize{
\item km: kmeans clustering object returned by stats::kmeans
\item geneClusters: a dataframe with columns: gene, cluster
}
}
\description{
This function perform k-means clustering on the profile matrix using stats::kmeans() function.
It tries to cluster the matrix by default "Hartigan-Wong" algorithm. If this algorithm fails
at Quick-TRANSfer stage, clustering is done with "MacQueen"
}
\examples{
NA
}
