% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_peak_annotation_file.R
\name{read_peak_annotation_file}
\alias{read_peak_annotation_file}
\title{Read peak target annotation file generated by bedtoolsClosestPeakAnnotate.pl}
\usage{
read_peak_annotation_file(title, file, cols)
}
\arguments{
\item{title}{Sample name to be used in the column names}

\item{file}{Peak annotation file}

\item{cols}{A vector of column names to select from the peak annotation file}
}
\value{
A dataframe with selected columns from peak annotation file
}
\description{
This function reads the peak target annotation file generated by bedtoolsClosestPeakAnnotate.pl.
It renames the peakEnrichment, peakPval and peakSummit columns as e.g. peakPval(SAMPLE_ID)
}
\examples{
NA
}
