% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{region_annotations}
\alias{region_annotations}
\title{Map peaks to given GRanges regions}
\usage{
region_annotations(peaksGr, featuresGr, includeFractionCut = 0.7, name = "tx")
}
\arguments{
\item{peaksGr}{GRanges object generated from narrowPeak or broadPeak file}

\item{featuresGr}{GRanges object for regions on which peaks needs to be mapped. E.g:
\code{GenomicFeatures::cdsBy()} or \code{GenomicFeatures::transcripts()}}

\item{includeFractionCut}{A floating point number between [0, 1]. If a peak covers more
than this proportion of feature, it is annotated as, eg. \code{include_tx}. Default: 0.7}

\item{name}{Feature type to be used as suffix in peak type annotation. Eg. CDS, gene etc.}
}
\value{
A modified peak GRanges object with additional columns: \code{tx_id,
peakType, targetOverlap, peakOverlap, peakDist, summitDist, relativePeakPos}
}
\description{
This function annotates the peaks onto a regions into e.g. \code{tx_start, tx_end,
include_tx, inside_tx} categories.
In addition, relativeSummitPos value is updated w.r.t. region for the peaks which are
annotated as e.g. \code{inside_tx}
}
\examples{
NA
}
