% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_functions.R
\name{region_coverage}
\alias{region_coverage}
\title{Calculate mean coverage for regions.}
\usage{
region_coverage(regions, format = "narrowPeak", bwFile, name = "coverage")
}
\arguments{
\item{regions}{Either GenomicRanges object or a file in bed, narrowPeak or broadPeak format}

\item{format}{format of the peak file. Should be one of "bed", "narrowPeak" or "broadPeak".
This option does not matter if the 'regions' argument if a GRanges object.}

\item{bwFile}{bigWig file from which the coverage will be calculated}

\item{name}{Name of the coverage column. Default: coverage}
}
\value{
If the regions is a GRanges object, then same object is returned with additional
coverage column. Otherwise a dataframe is returned.
}
\description{
This function calculate the coverage for the regions provided as bed/narrowPeak/broadPeak
format file from a bigWig file. For each region, it adds the depth and divide it by the
region length. NOTE: It assumes that the bigWig file is already scalled (e.g. 1 million)
and hence does perform scalling or normalization.
}
