% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_decoration_functions.R
\name{row_annotation_axis}
\alias{row_annotation_axis}
\title{Draw scale for heatmap annotation}
\usage{
row_annotation_axis(an, at, labels, slice)
}
\arguments{
\item{an}{name of the row annotation which needs to be modified}

\item{at}{a numeric vector x coordinates where the ticks and labels will be drawn. Eg.: c(0, 2000, 4000)}

\item{labels}{a vector of lables. Eg.: c("0kb", "2kb", ">4kb")}

\item{slice}{row slice}
}
\value{
The function returns no value.
}
\description{
Draw scale for heatmap annotation
}
\examples{
row_annotation_axis(c(0, 2000, 4000), c("0kb", "2kb", ">4kb"), "gene_length", 7)
}
