% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_scaling_functions.R
\name{rowScale}
\alias{rowScale}
\title{row wise z-score scaling of matrix}
\usage{
rowScale(x, center = TRUE, scale = TRUE, add_attr = TRUE,
  rows = NULL, cols = NULL)
}
\arguments{
\item{x}{A matrix}

\item{center}{Logical: whether to center each row by row mean or not.
Default: TRUE}

\item{scale}{Logical: scale each row by standard deviation or not. Default: TRUE}

\item{add_attr}{Logical: whether to add attributes to scalled matrix. Default: FALSE}

\item{rows}{Row indices if only a subset of rows need to be used. Default: NULL
i.e. all rows are used}

\item{cols}{Column indices if only a subset of columns needs to be used. Default: NULL
i.e. all columns are used}
}
\value{
A row scaled matrix
}
\description{
This function is similar to the \code{base::scale()}  function. This function is
copied from the tutorial posted by user \strong{strictlystat} on R-Bloggers.
Reference: \url{https://www.r-bloggers.com/a-faster-scale-function/}
}
\details{
IMP: Remember that default \code{base::scale()} function has different behaviour
when \code{center == FALSE}. If \code{center} is \code{TRUE}, the scaling is done
by dividing centered columns of x by standard deviation. If \code{center} is
\code{FALSE}, the root mean square of column. Refer to \code{\link[base]{scale}}
for details.
}
\examples{
NA
}
