% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{select_optimal_targets}
\alias{select_optimal_targets}
\title{Assign best target/s for each peak}
\usage{
select_optimal_targets(peakGr, insideSkewToEndCut = 0.7,
  promoterLength = 500, bindingInGene = FALSE)
}
\arguments{
\item{peakGr}{peak annotation in form of GRanges object. This is generated by
combining \code{UTR_annotate(), region_overlap_annotate(), upstream_annotate()}
functions.}

\item{insideSkewToEndCut}{A floating point number in range [0, 1]. If a peak is
present inside feature/gene and the relative summit position is > insideSkewToEndCut,
it is closer to the end of the feature. Default: 0.7}

\item{promoterLength}{Promoter length in number of nucleotides}

\item{bindingInGene}{Logical: whether the ChIPseq TF binds in gene body. This is
useful for polII ChIPseq data. Default: FALSE}
}
\value{
Same GRanges object as peakGr but with modified peakType column or excluding
some targets which were not optimum.
}
\description{
This function checks the different targets assigned for a peak and returns the
optimum target gene/s
}
\examples{
NA
}
