% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p06_signal_heatmap.R
\name{signal_heatmap}
\alias{signal_heatmap}
\title{Heatmap for polII binding signal}
\usage{
signal_heatmap(
  log2_matrix,
  htName = NULL,
  col_title = NULL,
  legend_title,
  color,
  htSplit = NULL,
  showLegend = TRUE,
  ...
)
}
\arguments{
\item{log2_matrix}{a log2 matrix with polII expression values to be plotted as Heatmap}

\item{htName}{heatmap name. Default: in not provided, col_title is used}

\item{col_title}{column title for heatmap. If not provided, matrix column names are used. Default: NULL}

\item{legend_title}{legend title}

\item{color}{color for heamap generated by ColorRamp2}

\item{htSplit}{if heatmap rows are to be split, the dataframe or vector with split information. Default: NULL}

\item{showLegend}{Binary: whether to show legend or not. Default: TRUE}

\item{...}{Other arguments to Heatmap function}
}
\value{
A complexheatmap object i.e. a heatmap which can be directly plotted or added to heatmap list
}
\description{
This function generate a polII binding signal heatmap using ComplexHeatmap package
function Heatmap().
}
\examples{
NA
}
