% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SM_genes_comparison_plot.R
\name{SM_genes_comparison_plot}
\alias{SM_genes_comparison_plot}
\title{Title}
\usage{
SM_genes_comparison_plot(data, orderFile, exptInfo, outFile, name, res)
}
\arguments{
\item{res}{}
}
\description{
Title
}
\examples{
NA
}
