% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{upstream_annotate}
\alias{upstream_annotate}
\title{Annotate upstream peaks on transcripts}
\usage{
upstream_annotate(peaksGr, featuresGr, txdb = NULL, txTypes = NULL,
  promoterLength, ...)
}
\arguments{
\item{peaksGr}{GRanges object for peak data}

\item{featuresGr}{A transcript GRanges object}

\item{txdb}{Optional TxDB object. A UTR less region is created using TxDB and used
to find the overlap between peak and its downstream target. A max fractional
overlap of 0.2 with geneA is allowed in a case when peak overlaps with a geneA and
is upstream of geneB. This is useful for the peaks which are near TES of a geneA.
If TxDB object is not provided, featuresGr is used. Default: featuresGr is used.}

\item{txTypes}{Types of transcripts to include from annotation. Should be a
character vector. This options is used only if TxDB object is provided. Default: NULL}

\item{promoterLength}{Promoter region length. Upstream peaks within \code{promoterLength}
distance of feature start are annotated as \code{promoter} region peaks.}

\item{...}{Other arguments for \code{nearest_upstream_bidirectional()} function}
}
\value{
A modified peak GRanges object with additional columns: \code{ tx_id,
peakType, featureCovFrac, peakDist, summitDist}
}
\description{
Annotate upstream peaks on transcripts
}
\examples{
NA
}
