% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{upstream_annotations}
\alias{upstream_annotations}
\title{Annotate upstream peaks on transcripts}
\usage{
upstream_annotations(
  peaksGr,
  featuresGr,
  txdb = NULL,
  upstreamOverlappingFraction = 0.2,
  promoterLength,
  upstreamLimit,
  bidirectionalDistance,
  bidirectionalSkew
)
}
\arguments{
\item{peaksGr}{GRanges object for peak data}

\item{featuresGr}{A transcript GRanges object}

\item{txdb}{Optional TxDB object. A 3UTR less region is created using TxDB and used
to find the overlap between peak and its downstream target. A max fractional
overlap of 0.2 with geneA is allowed in a case when peak overlaps with a geneA and
is upstream of geneB. This is useful for the peaks which are near TES of a geneA.
If TxDB object is not provided, featuresGr is used. Default: featuresGr is used.}

\item{upstreamOverlappingFraction}{See details. Default: 0.2}

\item{promoterLength}{Promoter region length. Upstream peaks within \code{promoterLength}
distance of feature start are annotated as \code{promoter} region peaks.}

\item{upstreamLimit}{Maximum distance of peak for upstream annotation. Peak beyond
this distance can be considered as intergenic instead.}

\item{bidirectionalDistance}{If a peak is present at bidirectional promoter where
distance between two TSS is < \code{bidirectionalDistance}, both the targets are
assigned to the peak. Default: 1000}

\item{bidirectionalSkew}{Maximum fraction of peak region allowed on the side of
false target from the midpoint of two target genes. Default: 0.2}
}
\value{
A modified peak GRanges object with additional columns: \code{ tx_id,
peakType, targetOverlap, peakOverlap, peakDist, summitDist, relativePeakPos}
}
\description{
This function annotates the peaks with nearest downstream target.
See \strong{Use of arguments} section for more details.
}
\section{Use of arguments}{

\subsection{upstreamOverlappingFraction}{
There will be cases when a peak is inside a gene and it is upstream of other gene.
\preformatted{
#                                                                            #
#           target1                     target2                              #
#         =====<=======<===       =====<=======<========<=======             #
#                                   -^--           --^-                      #
#                                 peak1           peak2                      #
#                         |<------>|                                         #
#                         |<------------------------->|                      #
#                                                                            #
In above cases, peak1 can be annotated as Upstream of target1. However not peak2
because target2 has bigger fraction in-between [target1, peak2] range

Target gene inside gene case:
#                                                                            #
#         =====<=======<=======<=======<========<======= target2             #
#          target1 ==<==                                                     #
#                                  -^--           --^-                      #
#                                 peak1           peak2                      #
#                       |<-------->|                                         #
#                                                                            #
}
In above case, peak1 is inside targe2 and upstream of target1. These targets are
selected in \code{select_optimal_targets()} if peak lies within promoter range.
}


\subsection{bidirectionalDistance and bidirectionalSkew}{
For peak at bidirectional promoter, its skewness from midpoint and summit poistion
w.r.t. midpoint is used to decide correct target/s. \cr
\code{nearest_upstream_bidirectional(..., bidirectionalDistance = 1000,
bidirectionalSkew = 0.2)}
\preformatted{
#                        *                                                   #
#                    |<-------more than 500bp------>|                        #
#         target1                    |                    target2            #
#  ==<=====<=====<===                |               ===>===>====>====>==    #
#                              --^---|- peak1                                #
#                                 ---|^--- peak2                             #
#                                  --|^------ peak3                          #
#                peak4  --^--        |                                       #
#                                    |                                       #
#                                  { | } central 10\% region                 #
#                                    |                                       #
#                        midpoint between two targets                        #
#                                                                            #
peak1 => target1: more than 80% of the peak lies on target1 side from TSS midpoint
         and peak summit is not in central 20\% region.
peak2 => target1, targe2: peak lies on the center
peak3 => Additionally, if peak summit (^) is within central 10\% region
         (= bidirectionalSkew/2), peak is assigned to both the target genes
peak4 => target1
If gap between two bidirectional genes < \code{bidirectionalDistance}
(default: 1000bp), peak is assigned to both the target genes
}
}
}

\examples{
NA
}
