% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{UTR_annotate}
\alias{UTR_annotate}
\title{Annotate peak targets in UTR regions}
\usage{
UTR_annotate(peaksGr, featuresGrl, utrType, txdb)
}
\arguments{
\item{peaksGr}{GRanges object generated from narrowPeak or broadPeak file}

\item{featuresGrl}{GRangesList object for CDS generated by \code{GenomicFeatures::fiveUTRsByTranscript()}
or \code{GenomicFeatures::threeUTRsByTranscript()}}

\item{utrType}{A character string either of \code{"5UTR"} or \code{"3UTR"}. This string is added
as peakType column value for the peaks which overlap the respective UTR region.}

\item{txdb}{TxDB object}
}
\value{
A modified peak GRanges object with additional columns: \code{ tx_id,
peakType, featureCovFrac, peakDist, summitDist}
}
\description{
Annotate peak targets in UTR regions
}
\examples{
NA
}
