% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{annotate_ranges}
\alias{annotate_ranges}
\title{Annotate GRanges using TxDB}
\usage{
annotate_ranges(peaks, txdb, promoterLength, txIds = NULL,
  blacklistRegions = NULL, excludeType = c("tRNA", "rRNA", "snRNA",
  "snoRNA", "ncRNA"), bidirectionalDistance = 500,
  bidirectionalSkew = 0.2, includeFractionCut = 0.7,
  bindingInGene = FALSE, insideSkewToEndCut = 0.7,
  removePseudo = FALSE)
}
\arguments{
\item{peaks}{A GRanges object with name column.}

\item{txdb}{TxDB object which will be used for annotation}

\item{promoterLength}{Promoter length in number of nucleotides.}

\item{txIds}{A vector of transcript IDs to be used specifically in the annotation
process instead of full transcript set. These should be internal tx_ids from TxDB
object. This is useful feature to exclude tRNA, rRNA transcripts while annotating
the regions. Default: NULL}

\item{blacklistRegions}{A BED file or GRanges object with ChIPseq blacklist regions.
Peaks overlapping with these regions are not used for annotation.}

\item{excludeType}{Types of transcripts to exclude from annotation. Should be a
character vector. Default: \code{c("tRNA", "rRNA", "snRNA", "snoRNA", "ncRNA")}}

\item{bidirectionalDistance}{If a peak is present at bidirectional promoter where
distance between two TSS is < bidirectionalDistance, both the targets are assigned
to the peak.}

\item{bidirectionalSkew}{Maximum fraction of peak region allowed on the side of
false target from the midpoint of two target genes. Default: 0.2}

\item{includeFractionCut}{Number between [0, 1]. If a peak covers more than this
fraction of feature/gene, it will be marked as include_tx/include_CDS. Default: 0.7}

\item{bindingInGene}{Logical: whether the ChIPseq TF binds in gene body. This is
useful for polII ChIPseq data. Default: FALSE}

\item{insideSkewToEndCut}{A floating point number in range [0, 1]. If a peak is
present inside feature/gene and the relative summit position is > insideSkewToEndCut,
it is closer to the end of the feature. Default: 0.7}

\item{removePseudo}{Logical: whether to remove peak targets which are marked as pseudo.
Default: FALSE}
}
\value{
A GenomicRanges object with peak annotation
}
\description{
This is a generic function to annotate GRanges object with TxDB annotations. \cr
Regions are annnotated with following \strong{broad categories} and \emph{specific
types} (listed in decreasing order of preference):
\enumerate{
\item \strong{featureInPeak:} \emph{"include_tx", "include_CDS"}
\item \strong{nearStart:} \emph{"5UTR", "CDS_start", "tx_start"}
\item \strong{nearEnd:} \emph{"3UTR", "tx_end", "CDS_end"}
\item \strong{peakInFeature:} \emph{"exon", "intron", "inside_tx", "inside_CDS"}
\item \strong{upstreamTss:} \emph{"promoter", "upstream"}
\item \strong{intergenic:} \emph{"intergenic"}
}
Additionally, a \emph{pseudo} prefix is added to the peakType where a peak is
annotated to two target genes/features and one of it is more optimum than other.
The less optimum target type is prefixed with \emph{pseudo}.\cr
See \strong{Use of arguments} section for more details.
}
\section{Use of arguments}{

\subsection{upstreamOverlappingFraction}{
There will be cases when a peak is inside a gene and it is upstream of other gene.
\preformatted{
#                                                                            #
#           target1                     target2                              #
#         =====<=======<===       =====<=======<========<=======             #
#                                   ---            ----                      #
#                                 peak1           peak2                      #
#                         |<------>|                                         #
#                         |<------------------------->|                      #
#                                                                            #
In above cases, peak1 can be annotated as Upstream of target1. However not peak2
because target2 has bigger fraction in-between [target1, peak2] range

Target gene inside gene case:
#                                                                            #
#         =====<=======<=======<=======<========<======= target2             #
#          target1 ==<==                                                     #
#                                   ---            ----                      #
#                                 peak1           peak2                      #
#                       |<-------->|                                         #
#                                                                            #
}
In above case, peak1 is inside targe2 and upstream of target1. These targets are
selected in \code{select_optimal_targets()} if peak lies within promoter range.
}


\subsection{bidirectionalDistance and bidirectionalSkew}{
For peak at bidirectional promoter, its skewness from midpoint and summit poistion
w.r.t. midpoint is used to decide correct target/s. \cr
\code{nearest_upstream_bidirectional(..., bidirectionalDistance = 500,
bidirectionalSkew = 0.2)}
\preformatted{
#                        *                                                   #
#                    |<-------more than 500bp------>|                        #
#         target1                    |                    target2            #
#  ==<=====<=====<===                |               ===>===>====>====>==    #
#                              --^---|- peak1                                #
#                                 ---|^--- peak2                             #
#                                    |                                       #
#                                  --|-^------ peak3                         #
#                                    |                                       #
#                                 {  |  } middle 20\% region                  #
#                                    |                                       #
#                        midpoint between two targets                        #
#                                                                            #
peak1 => target1: more than 80% of the peak lies on target1 side from TSS midpoint
         and peak summit is not in central 20\% region.
peak2 => target1, targe2: peak lies on the center
peak3 => Additionally, if peak summit (^) is within central 20\% region
         (bidirectionalSkew), peak is assigned to both the target genes
If gap between two bidirectional genes < bidirectionalDistance (default: 500bp),
peak is assigned to both the target genes
}
}


\subsection{insideSkewToEndCut}{
Use of \code{select_optimal_targets(..., insideSkewToEndCut = 0.7, promoterLength = 500)}
\preformatted{
#                                                                            #
#                 target1                *            target2                #
#         0    0.25     0.5     0.75     |<--500-->|                         #
#         =======>=======>=======>=======           =====>=====>===>===      #
#                                  ----                                      #
#                                  peak1                                     #
#         |<--------0.7------->|                                             #
#         **                                                                 #
#                                                                            #
In above example, peak1 is inside target1 but it is near the end
Relative position of the peak is >0.7 in target1. peak1 is also
upstream of target2 and within 500bp. So new annotation is
target1: pseudo_inside
target2: upstream
}
}
}

\examples{
NA
}
