% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_profile_matrix_import_functions.R
\name{bigwig_profile_matrix}
\alias{bigwig_profile_matrix}
\title{Generate profile matrix using normalizeToMatrix()}
\usage{
bigwig_profile_matrix(bwFile, regions, signalName, extend = c(2000,
  1000), w = 10, targetName = "gene", storeLocal = FALSE, localPath,
  ...)
}
\arguments{
\item{bwFile}{BigWig file for the signal}

\item{regions}{bed file for the regions/genes of interest or a GenomicRanges object}

\item{signalName}{Signal name}

\item{extend}{extend argument of normalizeToMatrix function. Default: c(2000, 1000)}

\item{w}{w argument of normalizeToMatrix function: Default: 10}

\item{targetName}{Name to be used for target. Eg: "gene", "TSS", "TES", "summit" etc. default: gene}

\item{storeLocal}{Logical. Whether to store the matrix locally as gzipped file.
Default: FALSE}

\item{localPath}{File path pointing to locally stored matrix}

\item{...}{other arguments passed to function \code{normalizeToMatrix}
function}
}
\value{
profile matrix of class normalizedMatrix
}
\description{
Generate profile matrix using normalizeToMatrix()
}
\examples{
NULL
}
