% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_matrix_import_functions.R
\name{bigwig_profile_matrix}
\alias{bigwig_profile_matrix}
\title{Generate profile matrix using normalizeToMatrix()}
\usage{
bigwig_profile_matrix(bwFile, bedFile, signalName, genes,
  readLocal = TRUE, storeLocal = FALSE, localPath, extend = c(2000,
  1000), target = "gene", w = 10, ...)
}
\arguments{
\item{bwFile}{BigWig file for the signal}

\item{bedFile}{bed file for the genome of interest}

\item{signalName}{Signal name}

\item{genes}{A vector of gene IDs which are to be plotted.. Only these genes' profile is
extracted from the profile matrix for plotting.}

\item{readLocal}{Logical. Whether to read the matrix from local path. If TRUE, file
provided by localPath is used to read the profile matrix. Default: TRUE}

\item{storeLocal}{Logical. Whether to store the matrix locally as gzipped file.
Default: FALSE}

\item{localPath}{File path pointing to locally stored matrix}

\item{extend}{extend argument of normalizeToMatrix function. Default: c(2000, 1000)}

\item{target}{One of "gene", "tss", "tes", "point". If "tss"/"tes"/"point", the region is
extened around single point i.e. TSS or TES or single point (e.g. summit). Otherwise,
the extension is done around the "gene" region}

\item{w}{w argument of normalizeToMatrix function: Default: 10}

\item{...}{other arguments passed to either `import_profile_from_file` or `normalizeToMatrix`
function}
}
\value{
profile matrix of class normalizedMatrix
}
\description{
Generate profile matrix using normalizeToMatrix()
}
\examples{
NULL
}
