% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p03_QC_replicate_correlation.R
\name{compare_replicates}
\alias{compare_replicates}
\title{Correlation plots for numeric data}
\usage{
compare_replicates(
  data,
  rep1Col,
  rep2Col,
  value = "FPKM",
  trans = "identity",
  pseudoCount = 0
)
}
\arguments{
\item{data}{A dataframe}

\item{rep1Col}{column name for replicate 1}

\item{rep2Col}{column name for replicate 1}

\item{value}{A character name for the value. Default: FPKM}

\item{trans}{trans argument from ggplot::scale_x_continuous() function.}

\item{pseudoCount}{A pseudo count to avoid log(0) = Inf error. All values less than
this number will be set to this number in XY scatter plot. Default: 0}
}
\value{
A list with following elements is returned.
\itemize{
\item \strong{data:} plot data used for plotting
\item \strong{figure:} A combined figure generated by \code{ggpubr::ggarrange}
\item \strong{plots:} Individual ggplot list for each plot.
\itemize{
\item \strong{table:} a \code{ggtable} object for quantile stats
\item \strong{density:} a FPKM density plot for both replicates
\item \strong{scatter:} A list with two scatter plot objects: value, rank
\item \strong{corrVariation:} A line plot showing changing correlation with quantile
subset of data.
}
}
}
\description{
Correlation plots for numeric data
}
\examples{
NA
}
