% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_fold_change_functions.R
\name{get_fold_change}
\alias{get_fold_change}
\title{Raw log2 fold change value between two samples}
\usage{
get_fold_change(df, s1, s2, newCol = "lfc", isExpressedCols = NULL)
}
\arguments{
\item{df}{data frame}

\item{s1}{column name for sample1}

\item{s2}{column name for sample2}

\item{newCol}{new column name for log2(fold_change) values. Default: lfc}

\item{isExpressedCols}{A named vector with column names for isExpressed data for
both polII samples}
}
\value{
dataframe with additional fold change column
}
\description{
Raw log2 fold change value between two samples
}
\examples{
NA
}
