% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess_functions.R
\name{import_peak_annotation}
\alias{import_peak_annotation}
\title{Import peak annotation data}
\usage{
import_peak_annotation(sampleId, peakAnnoFile, peakFile, bwFile,
  fcCutoff = 1, pvalCutoff = 1, columns = NULL)
}
\arguments{
\item{sampleId}{sample ID}

\item{peakAnnoFile}{macs2 peak annotation file generated by \code{narrowPeak_annotate()}.}

\item{peakFile}{macs2 output file. Should be narrowPeak or broadPeak file}

\item{bwFile}{bigWig format coverage track file}

\item{fcCutoff}{fold enrichment cutoff for the peak. Default: 1 i.e. no cutoff}

\item{pvalCutoff}{log10_pvalue cutoff for the peak. Default: 1 i.e. no cutoff}

\item{columns}{A vector of column names which should be returned. Allowed values are: \code{c(
"hasPeak", "peakPosition", "peakType", "peakId", "peakEnrichment", "peakPval", "peakQval",
"peakSummit", "peakDist", "summitDist", "bidirectional", "featureCovFrac", "relativeSummitPos",
"peakRegion", "peakCoverage")}. Default: all columns are returned}
}
\value{
A peak annotation as dataframe
}
\description{
This function reads the '*.nearestCDS.tab' file generated by function
\code{narrowPeak_annotate()}. For each peak, it calculates the peak coverage values.
}
\examples{
NA
}
