% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p03_data_preprocess_functions.R
\name{import_peak_annotation}
\alias{import_peak_annotation}
\title{Import peak annotation data}
\usage{
import_peak_annotation(
  sampleId,
  peakAnnoFile,
  removePseudo = TRUE,
  fcCutoff = 1,
  pvalCutoff = 1,
  columns = NULL,
  renameColumn = TRUE
)
}
\arguments{
\item{sampleId}{sample ID}

\item{peakAnnoFile}{macs2 peak annotation file generated by \code{narrowPeak_annotate()}.}

\item{removePseudo}{Logical: whether to remove peak targets which are marked as pseudo.
Default: TRUE}

\item{fcCutoff}{fold enrichment cutoff for the peak. Default: 1 i.e. no cutoff}

\item{pvalCutoff}{log10_pvalue cutoff for the peak. Default: 1 i.e. no cutoff}

\item{columns}{A vector of column names which should be returned. Allowed values are:
\code{c("peakPosition", "peakAnnotation", "peakId", "peakEnrichment", "peakPval", "peakQval",
"peakSummit", "peakDist", "summitDist", "bidirectional", "targetOverlap", "relativeSummitPos",
"peakRegion", "peakCoverage", "peakOverlap", "relativePeakPos")}. Default: all columns are returned}

\item{renameColumn}{Logical indicating whether to add sampleId suffix to column name. If
TRUE, each column name is suffixed with sampleId. Default: TRUE}
}
\value{
A peak annotation as dataframe
}
\description{
This function reads the '*.annotation.tab' file generated by function
\code{narrowPeak_annotate()}. Column names are prefixed with sampleId.
}
\examples{
NA
}
