% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_data_preprocess_functions.R
\name{import_peak_annotation}
\alias{import_peak_annotation}
\title{Import peak annotation data}
\usage{
import_peak_annotation(sampleId, peakAnnoFile, fcCutoff = 1,
  pvalCutoff = 1, columns = NULL)
}
\arguments{
\item{sampleId}{sample ID}

\item{peakAnnoFile}{macs2 peak annotation file generated by \code{narrowPeak_annotate()}.}

\item{fcCutoff}{fold enrichment cutoff for the peak. Default: 1 i.e. no cutoff}

\item{pvalCutoff}{log10_pvalue cutoff for the peak. Default: 1 i.e. no cutoff}

\item{columns}{A vector of column names which should be returned. Allowed values are:
\code{c("peakPosition", "peakType", "peakId", "peakEnrichment", "peakPval", "peakQval",
"peakSummit", "peakDist", "summitDist", "bidirectional", "featureCovFrac", "relativeSummitPos",
"peakRegion", "peakCoverage")}. Default: all columns are returned}
}
\value{
A peak annotation as dataframe
}
\description{
This function reads the '*.annotation.tab' file generated by function
\code{narrowPeak_annotate()}. Column names are prefixed with sampleId.
}
\examples{
NA
}
