% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_matrix_import_functions.R
\name{import_profile_from_file}
\alias{import_profile_from_file}
\title{Read profile matrix from file}
\usage{
import_profile_from_file(file, source = "deeptools", signalName, selectGenes,
  up = 200, target = 200, down = 100, binSize = 10,
  targetType = "gene", returnDf = FALSE)
}
\arguments{
\item{file}{profile matrix for e.g. generated by deeptools' "computeMatrix scale-regions"
command}

\item{source}{A character string for the source of the matrix being read. One of "deeptools",
"miao", "normalizedmatrix"}

\item{signalName}{name of the signal track}

\item{selectGenes}{A vector of gene IDs which are to be plotted.. Only these genes' profile is extracted
from the profile matrix for plotting.}

\item{up}{number of bins in upstream region}

\item{target}{number of bins in gene body}

\item{down}{number of bins in downstream region}

\item{binSize}{bin size used while generating profile matrix}

\item{targetType}{One of "gene", "TSS", "TES", "point". If target is "gene", target is used to decide
the number of bins in gene body. Otherwise, for TSS/TES/point, a signle point matrix is expected where
the region is extracted around single point. Default: gene}

\item{returnDf}{If TRUE, returns the dataframe instead of profile matrix. Default: FALSE}
}
\value{
profile matrix of class `normalizedMatrix`
}
\description{
This function reads the locally stored profile matrix and convert it
to `normalizedMatrix` class. This matrix can be used for generating profile plot
using EnrichedHeatmap package.
}
\examples{
NA
}
