% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot_list_functions.R
\name{multi_profile_plots}
\alias{multi_profile_plots}
\title{EnrichedHeatmap plots and ComplexHeatmaps for multiple samples}
\usage{
multi_profile_plots(exptInfo, genesToPlot, clusters = NULL,
  drawClusterAn = TRUE, clusterColor = NULL, clustOrd = NULL,
  profileColors = NULL, matSource = "deeptools", targetType = "gene",
  matBins = c(200, 200, 100, 10), ylimFraction = NULL,
  plotExpression = FALSE, expressionData = NULL, expressionColor = NULL,
  rasterize = TRUE, rasterQual = 5, ...)
}
\arguments{
\item{exptInfo}{experiment info as data frame with information like sampleID, type, path etc}

\item{genesToPlot}{A vector of gene IDs which are to be plotted}

\item{clusters}{Dataframe with cluster information. Two columns must be present in this
dataframe: 'cluster', 'gene'. Default: NULL}

\item{drawClusterAn}{Logical: Whether to add row annotation heatmap to heatmap list. Default: TRUE}

\item{clusterColor}{cluster annotation color information. Default: NULL}

\item{clustOrd}{A character vector of cluster order in the plot. If not provided,
clusters are arranged as per character sort order}

\item{profileColors}{a named list of color objects for each of the sample. Color for profile
heatmap should be generated by colorRamp2}

\item{matSource}{Method by which profile matrix was generated. One of "deeptools", "miao",
"normalizedmatrix"}

\item{targetType}{One of "gene", "TSS", "TES", "point". If target is "gene", target is used to decide
the number of bins in gene body. Otherwise, for TSS/TES/point, a signle point matrix is expected where
the region is extracted around single point. Default: gene}

\item{matBins}{A vector with four elements representing the column counts for up, body,
down regions and binSize for the profile matrix. Default: c(200, 200, 100, 10)}

\item{ylimFraction}{A named list with intensity scores to use as ylimit for top annotation.
If the value is single number, it has to be floating point number to extract the quantile and use
limit [0, quantile(x)]. If the value is numeric vector of length 2, the two numbers are used as lower
and upper limits for ylimit of top annotation. Default: NULL.}

\item{plotExpression}{Binry: whether to plot polII expression heatmap or not. Default: FALSE}

\item{expressionData}{an dataframe which has info: clustering, polII expression, TF binding
status for each gene etc. Default: NULL}

\item{expressionColor}{color object for polII expression heatmap}

\item{rasterize}{Binary: whether to rasterize the profile heatmap or not. By default this is set as TRUE
for multiple profile plots.}

\item{rasterQual}{Raster quality. This is the raster_quality argument Heatmap function. in Default: 5}

\item{...}{Other arguments for EnrichedHeatmap function}
}
\value{
A list object with following elements:
\itemize{
\item heatmapList: a heatmap list object which has profile heatmaps and optionally expression heatmaps
\item profileHeatmaps: a named list of profile heatmaps generated by EnrichedHeatmap
\item expressionHeatmaps: a named list of polII signal heatmaps
\item plotGaps: a vector with required gaps between heatmaps while generating multiprofile plots
\item profileColors: a named list object with color scale used for each profile heatmap
\item expressionColor: a named list object with color scale used for each expression heatmap
}
}
\description{
This method generates profile heatmaps for multiple samples. Optionally, if the sample list
has any polII samples, it also generate a heatmap for the polII expression values.
}
\examples{
NA
}
