% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{narrowPeak_annotate}
\alias{narrowPeak_annotate}
\title{Annotate narrowPeak using TxDB}
\usage{
narrowPeak_annotate(peakFile, txdb, includeFractionCut = 0.7,
  bindingInGene = FALSE, promoterLength = 500,
  insideSkewToEndCut = 0.7, bidirectionCut = 500,
  excludeType = c("tRNA", "rRNA", "snRNA", "snoRNA", "ncRNA"))
}
\arguments{
\item{peakFile}{A narroPeak file}

\item{txdb}{TxDB object which will be used for annotation}

\item{includeFractionCut}{Number between [0, 1]. If a peak covers more than this
fraction of feature/gene, it will be marked as include_tx/include_CDS. Default: 0.7}

\item{bindingInGene}{Logical: whether the ChIPseq TF binds in gene body. This is
useful for polII ChIPseq data. Default: FALSE}

\item{promoterLength}{Promoter length in number of nucleotides. Default: 500}

\item{insideSkewToEndCut}{A floating point number in range [0, 1]. If a peak is
present inside feature/gene and the relative summit position is > insideSkewToEndCut,
it is closer to the end of the feature. Default: 0.7}

\item{bidirectionCut}{Distance cutoff to decide a peak as bidirectional for two
target genes. Both the target genes should be withing these bp distance from peak.
Default: 500}

\item{excludeType}{Types of transcripts to exclude from annotation. Should be a
character vector. Default: \code{c("tRNA", "rRNA", "snRNA", "snoRNA", "ncRNA")}}
}
\value{
A GenomicRanges object with peak annotation
}
\description{
Annotate narrowPeak using TxDB
}
\examples{
NA
}
