% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{nearest_upstream_bidirectional}
\alias{nearest_upstream_bidirectional}
\title{True target for bidirectional peak}
\usage{
nearest_upstream_bidirectional(bdirTargets, skewFraction = 0.2,
  minTSS_gapForPseudo = 500)
}
\arguments{
\item{bdirTargets}{A dataframe with two rows for bidirectional targets}

\item{skewFraction}{Minimum fraction of peak region allowed on the side of false target
from the midpoint of two target genes. Default: 0.2}

\item{minTSS_gapForPseudo}{Valid distance between two target genes TSS to mark one as psuedo}
}
\value{
Same dataframe in which peakType for one of the target is marked as pseudo
}
\description{
This function uses following logic to select or reject target from bidirectional peak.
\preformatted{
use of minTSS_gapForPseudo (500)* and skewFraction (0.2)
#                      *                                                   #
#                     |<---more than 500bp--->|                            #
#       target1                    |                    target2            #
#   ==<=====<=====<===      peak1  |           ===>=====>=====>=====>==    #
#                          ------- | peak2                                 #
#                               ---|----                                   #
#                                  |                                       #
#                      midpoint between two targets                        #
#                                                                          #
peak1 => target1: more than 80% of the peak lies on target1 side
peak2 => target1, targe2: peak lies on the center
If gap between two bidirectional genes < minTSS_gapForPseudo (default: 500bp),
peak is assigned to both the targets
}
}
\examples{
NULL
}
