% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{nearest_upstream_bidirectional}
\alias{nearest_upstream_bidirectional}
\title{True target for bidirectional peak}
\usage{
nearest_upstream_bidirectional(targetDf, t1Idx, t2Idx,
  skewFraction = 0.2, minTSS_gapForPseudo = 500)
}
\arguments{
\item{targetDf}{A dataframe which has bidirectional targets for each peak.}

\item{t1Idx}{target1 index vector}

\item{t2Idx}{target2 index vector. IMP: \code{length(t1Idx)} should be equal to
\code{length(t2Idx)}}

\item{skewFraction}{Minimum fraction of peak region allowed on the side of false target
from the midpoint of two target genes. Default: 0.2}

\item{minTSS_gapForPseudo}{Valid distance between two target genes TSS to mark one as psuedo}
}
\value{
A vector of row index for pseudo targets.
}
\description{
This function uses following logic to select or reject target from bidirectional peak.
\cr \strong{Use of \code{minTSS_gapForPseudo} (Default: 500)* and \code{skewFraction}
(Default: 0.2):}
\preformatted{
#                      *                                                   #
#                     |<---more than 500bp--->|                            #
#       target1                    |                    target2            #
#   ==<=====<=====<===      peak1  |           ===>=====>=====>=====>==    #
#                          ------- | peak2                                 #
#                               ---|----                                   #
#                                  |                                       #
#                      midpoint between two targets                        #
#                                                                          #
peak1 => target1: more than 80% of the peak lies on target1 side
peak2 => target1, targe2: peak lies on the center
If gap between two bidirectional genes < minTSS_gapForPseudo (default: 500bp),
peak is assigned to both the targets
}
}
\examples{
NA
}
