% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_distribution_plot_functions.R
\name{plot_MA_gg}
\alias{plot_MA_gg}
\title{Generate MA plot}
\usage{
plot_MA_gg(df, s1, s2, title = NA, colorCol, pseudoCount = 0.1,
  ylim = NULL)
}
\arguments{
\item{df}{dataframe with values two columns for which MA plot is to be generated}

\item{s1}{sample 1 column name}

\item{s2}{sample 2 column name}

\item{title}{title to be showed on plot}

\item{colorCol}{column name which should be used for coloring the points}

\item{pseudoCount}{A small number to be added to the values to avoid log(0) error.
Default: 0.1}

\item{ylim}{A two element numeric vector for Y-limits of MA plot. Default: c(-4, 4)}
}
\value{
a ggplot2 object which has MA plot
}
\description{
This method generates MA plot using two numeric columns from dataframe
It can also color the genes based on the colorCol variable
}
\examples{
NA
}
