% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p10_distribution_plot_functions.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Generate XY scatter plot for two sets}
\usage{
plot_scatter(df, s1, s2, title = "Scatter plot", colorCol,
  transformation = "log2", pseudoCount = 0.1)
}
\arguments{
\item{df}{dataframe with values two columns for which scatter plot is to be generated}

\item{s1}{sample 1 column name}

\item{s2}{sample 2 column name}

\item{title}{title to be showed on plot}

\item{colorCol}{column name which should be used for coloring the points}

\item{transformation}{trans argument from ggplot::scale_x_continuous() function.
Default: log2}

\item{pseudoCount}{A small number to be added to the values to avoid log(0) error.
Default: 0.1}
}
\value{
a ggplot object
}
\description{
Generate XY scatter plot for two sets
}
\examples{
NA
}
