% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess_functions.R
\name{preProcess_macs2_results}
\alias{preProcess_macs2_results}
\title{Pre-process macs2 peak calling output}
\usage{
preProcess_macs2_results(title, peakAnnoFile, cdsFile, peakFile, bwFile,
  outFile, bindingInGene = FALSE, fcCutoff = 1, pvalCutoff = 1)
}
\arguments{
\item{title}{sample ID}

\item{peakAnnoFile}{macs2 peak target gene found by \code{macs2 narrowPeak output |
bedtools closest | perl bedtoolsClosestPeakAnnotate.pl}}

\item{cdsFile}{BED file with gene information}

\item{peakFile}{macs2 output file. Should be narrowPeak or broadPeak file}

\item{bwFile}{bigWig format coverage track file}

\item{outFile}{output file}

\item{bindingInGene}{whether the factor bind in gene body like polII or not}

\item{fcCutoff}{fold enrichment cutoff for the peak. Default: 1 i.e. no cutoff}

\item{pvalCutoff}{log10_pvalue cutoff for the peak. Default: 1 i.e. no cutoff}
}
\value{
Returns nothing
}
\description{
This function process the '*.nearestCDS.tab' file generated by command:
"macs2 narrowPeak output | bedtools closest | perl bedtoolsClosestPeakAnnotate.pl".
If a gene is target of multiple peaks, for TSS peaks preference order is: \emph{
includeFeature, overlapStart, insideOverlapStart, insideOverlapStartOverlapEnd,
inside, upstream, pseudo_upstream}
For TES peak, the preference order is: \emph{overlapEnd, pseudo_overlapEnd,
insideOverlapEnd, pseudo_insideOverlapEnd, pseudo_inside}
}
\examples{
NA
}
