% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess_functions.R
\name{preProcess_polII_expression}
\alias{preProcess_polII_expression}
\title{Pre-process polII expression data}
\usage{
preProcess_polII_expression(expMat, title, expFraction, polIIExpFile)
}
\arguments{
\item{expMat}{Path to the polII expression file generated by Miao's script}

\item{title}{Sample id to be used}

\item{expFraction}{Fraction of the genes to be considered as expressed
(normally 10\% genes are under active transcription in the cell)}

\item{polIIExpFile}{Complete path for processed output file}
}
\value{
complete output file path
}
\description{
Pre-process polII expression data
}
\examples{
NA
}
