% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p06_profile_heatmap.R
\name{profile_heatmap}
\alias{profile_heatmap}
\title{Generate profile heatmap}
\usage{
profile_heatmap(
  profileMat,
  signalName,
  profileColor,
  columnTitle = NULL,
  createClusters = FALSE,
  numClust = 7,
  geneGroups = NULL,
  showAnnotation = TRUE,
  clusterColor = NULL,
  clusterOrder = NULL,
  ylimFraction = NULL,
  posLineGpar = gpar(lty = 2, alpha = 0.4, lwd = 0.5),
  rasterPar = list(use = TRUE, qual = 5),
  ...
)
}
\arguments{
\item{profileMat}{profile matrix of class \code{normalizedMatrix}}

\item{signalName}{name of the signal track}

\item{profileColor}{Color function for profile heatmap generated by \code{colorRamp2}}

\item{columnTitle}{title to be placed on the column. Default: signalName}

\item{createClusters}{Logical: whether or not to do kmeans clustering. If FALSE,
clusterStorePath must be provided which has the cluster data. Default: FALSE}

\item{numClust}{number of clusters to generate using k-means clustering. Default: 7}

\item{geneGroups}{Either a dataframe or a file path with cluster information or NULL.
Two columns must be present in this dataframe: \code{geneId, cluster}. In case it is NULL, heatmap
rows are not split into different groups.}

\item{showAnnotation}{Logical: to show or hide top annotation. Default: TRUE}

\item{clusterColor}{Color assignment object for cluster colors. If nothing provided, colors
are generated internally. Default: NULL}

\item{clusterOrder}{A character vector of cluster order in the plot. If not provided,
clusters are arranged as per character sort order}

\item{ylimFraction}{A numeric vector with one or two elements for deciding the \code{ylimit} of top
profile heatmap annotation. If the value is single number, it has to be floating point number
to extract the quantile and use limit \code{[0, quantile(x)]}. If the value is numeric vector of
length 2, the two numbers are used as lower and upper limits for ylimit of top annotation.
Default: NULL.}

\item{posLineGpar}{\code{pos_line_gp} parameter of \code{EnrichedHeatmap}. Default:
\code{gpar(lty = 2, alpha = 0.4, lwd = 0.5)}}

\item{rasterPar}{rasterization parameters as list() with two elements: use, qual. Default:
\code{list(use = TRUE, qual = 5)}. \code{rasterPar$use} is used for \code{use_raster} argument and
\code{rasterPar$qual} is used for \code{raster_quality} argument of \code{Heatmap} function.}

\item{...}{Other arguments for EnrichedHeatmap function}
}
\value{
A list object with following elements:
\itemize{
\item heatmap: Enriched heatmap
\item rowGroupHt: row group heatmap
\item legend: Annotation legend for clusters
\item cluster: dataframe with cluster information
\item clusterColor: cluster color assignment information
\item profileColor: Heatmap profile color function generated by colorRamp2
\item ylim: ylimit used for the top annotation Y axis limit
\item kmeans: if kmeans clustering was performed, a kmeans clustering result object
}
}
\description{
This function generate a profile heatmap using EnrichedHeatmap function. The genes
are clustered into different groups based on the profile using k-means clustering
algorithm. Optionally, the prior clustering information can be provided through
argument geneGroups
}
\examples{
NA

}
