% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_heatmap.R
\name{profile_heatmap}
\alias{profile_heatmap}
\title{Generate profile heatmap}
\usage{
profile_heatmap(profileMat, signalName, columnTitle = NULL,
  createClusters = FALSE, numClust = 7, geneGroups = NULL, profileColor,
  clusterColor = NULL, clusterOrder = NULL, ylimFraction = NULL, ...)
}
\arguments{
\item{profileMat}{profile matrix of class normalizedMatrix}

\item{signalName}{name of the signal track}

\item{columnTitle}{title to be placed on the column. Default: signalName}

\item{createClusters}{Binary: whether or not to do kmeans clustering. If FALSE,
clusterStorePath must be provided which has the cluster data. Default: FALSE}

\item{numClust}{number of clusters to generate using k-means clustering. Default: 7}

\item{geneGroups}{Either a dataframe or a file path with cluster information or NULL.
Two columns must be present in this dataframe: 'gene', 'cluster'. In case it is NULL, heatmap
rows are not split into different groups.}

\item{profileColor}{Color function for profile heatmap generated by colorRamp2}

\item{clusterColor}{Color assignment object for cluster colors. If nothing provided, colors
are generated internally. Default: NULL}

\item{clusterOrder}{A character vector of cluster order in the plot. If not provided,
clusters are arranged as per character sort order}

\item{ylimFraction}{A numeric vector with one or two elements for deciding the ylimit of top
profile heatmap annotation. If the value is single number, it has to be floating point number
to extract the quantile and use limit [0, quantile(x)]. If the value is numeric vector of
length 2, the two numbers are used as lower and upper limits for ylimit of top annotation.
Default: NULL.}

\item{...}{Other arguments for EnrichedHeatmap function}
}
\description{
This function generate a profile heatmap using EnrichedHeatmap function. The genes
are clustered into different groups based on the profile using k-means clustering
algorithm. Optionally, the prior clustering information can be provided through
argument geneGroups
}
\examples{
NA

}
