% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_matrix_import_functions.R
\name{profile_matrix_list}
\alias{profile_matrix_list}
\title{Generate profile matrix list}
\usage{
profile_matrix_list(exptInfo, geneList, ...)
}
\arguments{
\item{exptInfo}{experiment info as data frame with information like sampleID, type, path etc}

\item{geneList}{A vector of gene IDs for which the profile needs to be extracted}

\item{...}{Other argument to `import_profile_from_file` function}
}
\value{
A list of normalized matrix where each matrix is of normalizedMatrix class
}
\description{
This function reads/generate the profile matrix of normalizedMatrix class for each
sample
}
\examples{
NA
}
