% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{select_optimal_targets}
\alias{select_optimal_targets}
\title{Assign best target/s for each peak}
\usage{
select_optimal_targets(
  targetGr,
  promoterLength,
  upstreamLimit,
  bindingInGene,
  insideSkewToEndCut,
  bidirectionalSkew,
  bidirectionalDistance
)
}
\arguments{
\item{targetGr}{peak annotation in form of GRanges object. This is generated by
combining \code{UTR_annotate(), region_annotations(), upstream_annotations()}
functions.}

\item{promoterLength}{Promoter region length. Upstream peaks within \code{promoterLength}
distance of feature start are annotated as \code{promoter} region peaks.}

\item{upstreamLimit}{Maximum distance of peak for upstream annotation. Peak beyond
this distance can be considered as intergenic instead.}

\item{bindingInGene}{Logical: whether the ChIPseq TF binds in gene body. This is
useful for polII ChIPseq data. Default: FALSE}

\item{insideSkewToEndCut}{A floating point number in range [0, 1]. If a peak is
present inside feature/gene and the relative summit position is >
\code{insideSkewToEndCut}, it is closer to the end of the feature. Default: 0.7}

\item{bidirectionalSkew}{Maximum fraction of peak region allowed on the side of
false target from the midpoint of two target genes. Default: 0.2}

\item{bidirectionalDistance}{When a peak is present at bidirectional promoter where
distance between two TSS is < \code{bidirectionalDistance}, any gene within
\code{promoterLength} distance of peak is assigned to the peak as annotation. Default: 1000}
}
\value{
Same GRanges object as peakGr but with modified peakType column or excluding
some targets which were not optimum.
}
\description{
This function checks the different targets assigned for a peak and returns the
optimum target gene/s. See \strong{Use of arguments} section for more details.
}
\section{Use of arguments}{

\subsection{insideSkewToEndCut}{
Use of \code{select_optimal_targets(..., insideSkewToEndCut = 0.7, promoterLength = 500)}
\preformatted{
#                                                                            #
#                  d<500bp                                                   #
#      target1    |<---->|    target2           |<--500-->|      target3     #
#                        0  0.25  0.5  0.75  1                               #
#    ====<====<===       =====>=====>=====>=~~~~           ====>====>===     #
#                         --^--            --^--                             #
#                         peak1            peak2                             #
#                        |<----0.7---->|                                     #
#                        **                                                  #
#                                                                            #
In first example, peak1 is inside target2 and it is near the start of target2.
Even though peak1-target1 distance is < 500bp, target1 is marked as pseudo here.
In another example, peak2 is inside target2 but it is near the end at 3'UTR.
Relative position of the peak2 is >0.7 in target2. peak2 is also upstream of
target3 and within (upstreamLimit)bp. So correct annotations are:
peak1: target2
peak2: target2, target3
}
}
}

\examples{
NA
}
