% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{select_optimal_targets}
\alias{select_optimal_targets}
\title{Assign best target/s for each peak}
\usage{
select_optimal_targets(targetGr, promoterLength, bindingInGene,
  insideSkewToEndCut)
}
\arguments{
\item{targetGr}{peak annotation in form of GRanges object. This is generated by
combining \code{UTR_annotate(), region_annotations(), upstream_annotations()}
functions.}

\item{promoterLength}{Promoter length in number of nucleotides}

\item{bindingInGene}{Logical: whether the ChIPseq TF binds in gene body. This is
useful for polII ChIPseq data. Default: FALSE}

\item{insideSkewToEndCut}{A floating point number in range [0, 1]. If a peak is
present inside feature/gene and the relative summit position is > insideSkewToEndCut,
it is closer to the end of the feature. Default: 0.7}
}
\value{
Same GRanges object as peakGr but with modified peakType column or excluding
some targets which were not optimum.
}
\description{
This function checks the different targets assigned for a peak and returns the
optimum target gene/s.
}
\details{
\strong{Use of \code{insideSkewToEndCut} (0.7)* and \code{promoterLength} (500)**: }
\preformatted{
#                                                                #
#           target1                *            target2          #
#   0    0.25     0.5     0.75     |<--500-->|                   #
#   =======>=======>=======>=======         =====>=====>===      #
#                            ----                                #
#                            peak1                               #
#   |<--------0.7------->|                                       #
#   **                                                           #
#                                                                #
In above example, peak1 is inside target1 but it is near the end
Relative position of the peak is >0.7 in target1. peak1 is also
upstream of target2 and within 500bp. So new annotation is
target1: pasudo_inside
target2: upstream
}
}
\examples{
NA
}
