% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{set_peakTarget_to_pseudo}
\alias{set_peakTarget_to_pseudo}
\title{Set peakType to pseudo}
\usage{
set_peakTarget_to_pseudo(target)
}
\arguments{
\item{target}{A dataframe or GRanges object which has peakType column}
}
\value{
Same object with \code{pseudo} prefix to peakType column values
}
\description{
Set peakType to pseudo
}
\examples{
NA
}
