% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_peak_annotation.R
\name{splicing_unit_annotate}
\alias{splicing_unit_annotate}
\title{Annotate peaks with 5UTR/exon/intron/3UTR overlap}
\usage{
splicing_unit_annotate(peaksGr, featuresGr, featureType, txdb)
}
\arguments{
\item{peaksGr}{GRanges object for peak data}

\item{featuresGr}{A transcript GRanges object}

\item{featureType}{Feature type. One of \code{c("5UTR", "exon", "intron", "3UTR")}}

\item{txdb}{TxDB object. Transcript information for the exon/intron overlapping with peak
is extracted from TxDB object and used to calculate other features. In case a peak
overlap with exons/introns from multiple transcripts of same gene, longest transcript is
selected.}
}
\value{
A modified peak GRanges object with additional columns: \code{tx_id,
peakType, featureCovFrac, peakDist, summitDist}
}
\description{
Annotate peaks with 5UTR/exon/intron/3UTR overlap
}
\examples{
NA
}
