% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p03_QC_replicate_correlation.R
\name{tf_replicate_plots}
\alias{tf_replicate_plots}
\title{TF ChIPseq replicate correlation plots}
\usage{
tf_replicate_plots(sampleInfo, compare = "pvalue", title)
}
\arguments{
\item{sampleInfo}{sample information dataframe}

\item{compare}{One of \emph{pvalue} or \emph{enrichment}}

\item{title}{Comparison title}
}
\value{
A list with following ggplot2 elements:
\itemize{
\item \strong{table:} peak overlap summary table for replicates as a \code{ggtable} object
\item \strong{distribution:} Beeswarm plots showing comparative distribution between
replicates for a metric of choice
\item \strong{valueScatter:} X-Y scatter plot using values a metric of choice
\item \strong{rankScatter:} X-Y scatter plot using \code{rank(values)} a metric of choice
}
}
\description{
This function generate the peak overlap statistics table for two TF ChIPseq
replicates. Using common peaks, it plots the distribution and scatter plot of
either -log10(p-value) or fold enrichment from peak file. Macs2 peak caller
output files in narrowPeak/broadPeak format are used to extract the values.
}
\examples{
NA
}
