% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{upstream_annotate}
\alias{upstream_annotate}
\title{Annotate upstream peaks on transcripts}
\usage{
upstream_annotate(peaksGr, featuresGr, txdb, ...)
}
\arguments{
\item{peaksGr}{GRanges object for peak data}

\item{featuresGr}{A transcript GRanges object}

\item{...}{Other arguments for \code{nearest_upstream_bidirectional()} function}
}
\value{
A modified peak GRanges object with two additional columns: \code{peakType, tx_id}
}
\description{
Known issue: upstream targets will not be found for a peak which is inside
the CDS of a gene. It gets filtered when number of CDS overlapping in the
peak-target gap region are counted. By default current overlapping CDS is
 counted and target is filtered.
}
\examples{
NA
}
