% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p02_peak_annotation.R
\name{upstream_annotate}
\alias{upstream_annotate}
\title{Annotate upstream peaks on transcripts}
\usage{
upstream_annotate(peaksGr, featuresGr, txdb = NULL, excludeType = NULL,
  upstreamOverlappingFraction = 0.2, promoterLength, ...)
}
\arguments{
\item{peaksGr}{GRanges object for peak data}

\item{featuresGr}{A transcript GRanges object}

\item{txdb}{Optional TxDB object. A UTR less region is created using TxDB and used
to find the overlap between peak and its downstream target. A max fractional
overlap of 0.2 with geneA is allowed in a case when peak overlaps with a geneA and
is upstream of geneB. This is useful for the peaks which are near TES of a geneA.
If TxDB object is not provided, featuresGr is used. Default: featuresGr is used.}

\item{excludeType}{tx types to exclude from TxDB}

\item{upstreamOverlappingFraction}{See details. Default: 0.2}

\item{promoterLength}{Promoter region length. Upstream peaks within \code{promoterLength}
distance of feature start are annotated as \code{promoter} region peaks.}

\item{...}{Other arguments for \code{nearest_upstream_bidirectional()} function}
}
\value{
A modified peak GRanges object with additional columns: \code{ tx_id,
peakType, targetOverlap, peakOverlap, peakDist, summitDist, relativePeakPos}
}
\description{
This function annotates the peaks with nearest downstream target. See Details.
}
\details{
There will be cases when a peak is inside a gene and it is upstream of other gene.
\cr \strong{Use of \code{upstreamOverlappingFraction} (default: 0.2): }
\preformatted{
#                                                                         #
#        target1                     target2                              #
#      =====<=======<===       =====<=======<========<=======             #
#                                ---            ----                      #
#                              peak1           peak2                      #
#                      |<------>|                                         #
#                      |<------------------------->|                      #
#                                                                         #
in above cases, peak1 can be annotated as Upstream of target1. However not peak2
because target2 has bigger fraction in-between [target1, peak2] range

Target gene inside gene case:
#                                                                         #
#              target1          target2 (longer)                          #
#      =====<=======<=======<=======<========<=======                     #
#               ==<==                                                     #
#                                ---            ----                      #
#                              peak1           peak2                      #
#                    |<-------->|                                         #
#                                                                         #
}
In above case, peak1 is inside targe2 and upstream of target1. These targets are
selected in \code{select_optimal_targets()} if peak lies within promoter range.
}
\examples{
NA
}
