% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{upstream_annotate}
\alias{upstream_annotate}
\title{Annotate upstream peaks on transcripts}
\usage{
upstream_annotate(peaksGr, featuresGr, txdb = NULL, ...)
}
\arguments{
\item{peaksGr}{GRanges object for peak data}

\item{featuresGr}{A transcript GRanges object}

\item{txdb}{Optional TxDB object. A UTR less region is created using TxDB and used
to find the overlap between peak and its downstream target. A max fractional
overlap of 0.2 with geneA is allowed in a case when peak overlaps with a geneA and
is upstream of geneB. This is useful for the peaks which are near TES of a geneA.
If TxDB object is not provided, featuresGr is used. Default: featuresGr is used.}

\item{...}{Other arguments for \code{nearest_upstream_bidirectional()} function}
}
\value{
A modified peak GRanges object with additional columns: \code{ tx_id,
peakType, featureCovFrac, peakDist, summitDist}
}
\description{
Annotate upstream peaks on transcripts
}
\examples{
NA
}
